Moved char_array.py to char.py
