+++ data_highlight/highlighter.py
@@ -1,4 +1,4 @@
-from .support.char_array import Char
+from .support.char import Char
 from .support.line import Line
 from .highlighter_functionality.export import export_report
 from .support.token import SubToken
