+++ data_highlight/support/char.py
